MAF-17928: fix test code

+++ tests/test_flash_attn_ck.py
@@ -1353,7 +1353,7 @@ def test_flash_attn_race_condition(seqlen_q, seqlen_k, d, dropout_p, causal, dty
             assert dq_equal
 
 
-@pytest.mark.parametrize("dtype", [torch.float16])
+@pytest.mark.parametrize("dtype", [torch.bfloat16])
 @pytest.mark.parametrize("causal", [False, True])
 @pytest.mark.parametrize("d", [16, 32, 64])
 @pytest.mark.parametrize("seqlen", [1, 2, 5, 17, 128])
@@ -1461,7 +1461,7 @@ def test_flash_attn_bwd_transpose(seqlen, d, causal, dtype):
     ).abs().max().item()
 
 
-@pytest.mark.parametrize("dtype", [torch.float16])
+@pytest.mark.parametrize("dtype", [torch.bfloat16])
 @pytest.mark.parametrize("causal", [False, True])
 @pytest.mark.parametrize("d", [16, 32, 64])
 def test_flash_attn_bwd_varlen_overflow(d, causal, dtype):
